chore: rework StatusPanel and add status as a tab
