+++ src/locales/en.json
@@ -560,15 +560,12 @@
             },
             "Filament": "Filament",
             "File": "File",
-            "Files": {
-                "EmptyGcodes": "No G-Code available.",
-                "EmptyJobqueue": "There is currently no file in the job queue.",
-                "Filament": "Filament",
-                "Files": "Files",
-                "Jobqueue": "Jobqueue ({count})",
-                "JobqueueMoreFiles": "no more jobs | one more job | {count} more jobs",
-                "PrintTime": "Print Time"
-            },
+            "Files": "Files",
+            "EmptyGcodes": "No G-Code available.",
+            "EmptyJobqueue": "There is currently no file in the job queue.",
+            "Jobqueue": "Jobqueue ({count})",
+            "JobqueueMoreFiles": "no more jobs | one more job | {count} more jobs",
+            "PrintTime": "Print Time",
             "Flow": "Flow",
             "Headline": "Status",
             "Layer": "Layer",
@@ -581,6 +578,7 @@
             "ResumePrint": "Resume print",
             "Slicer": "Slicer",
             "Speed": "Speed",
+            "Status": "Status",
             "Total": "Total",
             "Unknown": "Unknown"
         },
